moved segment rows to hook fixed final segment not being locked to the bottom
